Use the single-transaction option with MySQL database dumps
Without this flag being set we can hit issues with certain
databases like:

SELECT, LOCK TABLES command denied to user u for table t when using LOCK TABLES

This either requires more permissions to be given to the user, or
for this flag to be set. So this is the least intrusive option.