INTLY-1100 Add monitoring label and concurrency policy
Allow the concurrency policy to be overridden and default the value
to Forbid. The reasoning for this is to prevent multiple occurrences
of the same Job to run at the same time.

Add monitoring labels to the CronJob and the Jobs that it creates
so that monitoring alerts can reference that when referring to
all Integreatly CronJobs instead of just having a static list.

Verification:
- Modify the cron schedule in the CronJob template to be something
like `*/1 * * * *`, to run once per minute.
- Do an `oc create` on the template.
- Ensure that the created CronJob has the `concurrencyPolicy` of
`Forbid`.
- Ensure that the created CronJob has a label with the name
`monitoring-key` and the value `middleware`.
- Ensure that a Job created by the CronJob has a label with the
name `monitoring-key` and the value `middleware`.