Reduce Retries
The number of retries was too high for what might be happening. When
dealing with a large number of files the chance for the a "file
changed as we read it" error grow. Rerunning oc cp increases the run
time for the backup pushing it closer to the 10 minute alert mark.